Create Package Infrastructure for Stan Package

Create directory and helper files for a Stan package using CmdStanR

setup_stan_package(loc = NULL, use_all = TRUE)

Arguments

loc
a file.path indicating package root with a default of the current working directory

use_all
a logical indicating if each section of Stan code should have it's own directory with a default of TRUE

Value

invisible null

Examples

# Create Temporary Directory
temp_dir <- tempdir()

# Now setup the structur
setup_stan_package(loc = temp_dir)
#> 
#> Creating directory structure
#> ℹ Place your Stan programs in the `inst/stan` directory
#> 
#> To reference Stan code in expanded directories use `#include
#> functions/func.stan syntax in the appropriate block`

# Examine output

list.files(temp_dir, recursive = TRUE)
#>  [1] "downlit/base"                                    
#>  [2] "downlit/remotes"                                 
#>  [3] "downlit/utils"                                   
#>  [4] "file2c7765a22c67/404.html"                       
#>  [5] "file2c7765a22c67/authors.html"                   
#>  [6] "file2c7765a22c67/bootstrap-toc.css"              
#>  [7] "file2c7765a22c67/bootstrap-toc.js"               
#>  [8] "file2c7765a22c67/CODE_OF_CONDUCT.html"           
#>  [9] "file2c7765a22c67/docsearch.css"                  
#> [10] "file2c7765a22c67/docsearch.js"                   
#> [11] "file2c7765a22c67/index.html"                     
#> [12] "file2c7765a22c67/LICENSE-text.html"              
#> [13] "file2c7765a22c67/LICENSE.html"                   
#> [14] "file2c7765a22c67/link.svg"                       
#> [15] "file2c7765a22c67/pkgdown.css"                    
#> [16] "file2c7765a22c67/pkgdown.js"                     
#> [17] "file2c7765a22c67/pkgdown.yml"                    
#> [18] "file2c7765a22c67/reference/clear_stan_cache.html"
#> [19] "file2c7765a22c67/reference/copy_models.html"     
#> [20] "file2c7765a22c67/reference/find_stan_code.html"  
#> [21] "file2c7765a22c67/reference/index.html"           
#> [22] "file2c7765a22c67/reference/Rplot001.png"         
#> [23] "file2c776b878a4e"                                
#> [24] "functions/func.stan"                             
#> [25] "inst/stan/base.stan"                             
#> [26] "libloc_177_22be9ae6e263548.rds"                  
#> [27] "libloc_182_d3198563f59e5c50.rds"                 
#> [28] "R/staninside-helpers.R"                          
#> [29] "test.stan"                                       
#> [30] "test2.stan"