Create Package Infrastructure for Stan Package

Create directory and helper files for a Stan package using CmdStanR

setup_stan_package(loc = NULL, use_all = TRUE)

Arguments

loc

a file.path indicating package root with a default of the current working directory

use_all

a logical indicating if each section of Stan code should have it's own directory with a default of TRUE

Value

invisible null

Examples

# Create Temporary Directory
temp_dir <- tempdir()

# Now setup the structur
setup_stan_package(loc = temp_dir)
#> Creating directory structure
#> ℹ Place your Stan programs in the `inst/stan` directory
#> To reference Stan code in expanded directories use `#include
#> functions/func.stan syntax in the appropriate block`

# Examine output

list.files(temp_dir, recursive = TRUE)
#>  [1] "R/staninside-helpers.R"          "downlit/base"                   
#>  [3] "downlit/remotes"                 "downlit/utils"                  
#>  [5] "file9f5eb969182"                 "functions/func.stan"            
#>  [7] "inst/stan/base.stan"             "libloc_193_8482e75eead890f7.rds"
#>  [9] "libloc_219_8cfeeab4a45b6595.rds" "libloc_243_4d8080fc42f8a239.rds"
#> [11] "test.stan"                       "test2.stan"